+++ components/layout/FAQ/index.tsx
